refactor: move watch into its own package
This relocates the majority of code for the factory watch command into
its own subdirectory/package.
